Разделил тесты по объектам (user, resource)
